+++ Config/ProjectSettings.ps1
@@ -419,9 +419,6 @@ if (!(Get-Variable -Name CheckProjectConstants -Scope Global -ErrorAction Ignore
 	# Project version, does not apply to non migrated 3rd party modules which follow their own version increment, do not modify!
 	New-Variable -Name ProjectVersion -Scope Global -Option Constant -Value ([version]::new(0, 9, 0))
 
-	# Default header that will appear in log files
-	New-Variable -Name DefaultLogHeader -Scope Global -Option Constant -Value "Administrator logs"
-
 	# Required minimum operating system version (v1809)
 	# TODO: v1809 needs to be replaced with minimum v1903, downgraded here because of Server 2019
 	# https://docs.microsoft.com/en-us/windows/release-information