+++ Readme/CHANGELOG.md
@@ -77,7 +77,7 @@ If you would like to see fresh changes done since last release you can do so on 
 
 - Test
 
-  - Unit tests will write logs to separate folder (In progress...)
+  - Unit tests will write logs to separate directory
 
 - Bugfix